requirements.txt
flask
flask-login
python-dotenv
werkzeug
itsdangerous
psycopg2-binary
